Refactor code
- Split up key_functions into multiple modules
